Fix some bugs with displaying homepage and added a few extra jte files for later changes

+++ src/main/java/cs3220/ai_workout_generator/WorkoutController.java
@@ -87,7 +87,7 @@ public class WorkoutController {
     // GET /workouts/{id} → view one
     // -------------------------------
     @GetMapping("/workouts/{id}")
-    public String viewWorkout(@PathVariable Long id,
+    public String viewWorkout(@PathVariable int id,
                               Model model,
                               HttpSession session) {
